pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lisa-grape-pipeline"
version = "0.1.0"
description = "Hyperspectral grape monitoring pipeline and pretrained models"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Ciem Cornelissen"}]

dependencies = [
    "h5py==3.10.0",
    "matplotlib",
    "memory_profiler==0.61.0",
    "numpy<2.0",
    "opencv_python==4.10.0.84",
    "pandas",
    "Pillow==11.3.0",
    "plotly==6.3.1",
    "pytz==2024.1",
    "scikit_learn==1.7.2",
    "scipy==1.16.2",
    "spectral==0.23.1",
    "toml==0.10.2",
    "torch==2.2.1",
    "torcheval==0.0.7",
    "torchmetrics==1.8.2",
    "tqdm==4.66.2",
    "ultralytics==8.3.213",
    "wandb==0.22.2",
]

[project.scripts]
lisa = "lisa.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
